de-ass/requirements.txt
kafka-python==2.2.2
requests==2.32.4
pyspark==3.5.1
PyYAML==6.0.2
pymongo==4.7.3
numpy<2
matplotlib<3.9
pandas>=1.5,<2.2
rich>=13.7
neo4j>=5,<6
streamlit>=1.36
plotly>=5
pyarrow>=12
